requirements.txt
Django==3.2.18
django-database-url==1.0.3
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
asgiref==3.6.0
sqlparse==0.4.3
django-cors-headers==3.14.0